remove `registerCustomFunction`: define functions where they are being used
