+++ app/src/instrumentation.ts
@@ -1,7 +1,6 @@
 import chalk from 'chalk'
 import { exportApiIdentifier } from './app/regionen/[regionSlug]/_mapData/mapDataSources/export/exportIdentifier'
 import { interactivityConfiguration } from './app/regionen/[regionSlug]/_mapData/mapDataSources/generalization/interacitvityConfiguartion'
-import { registerCustomFunctions } from './instrumentation/registerCustomFunctions'
 import { registerExportFunctions } from './instrumentation/registerExportFunctions'
 import { registerGeneralizationFunctions } from './instrumentation/registerGeneralizationFunctions'
 import { runAnalysis } from './instrumentation/runAnalysis'
@@ -10,8 +9,6 @@ export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
     try {
       const greenHook = chalk.bold(chalk.green(' ✓'))
-      await registerCustomFunctions()
-      console.log(greenHook, 'Custom SQL functions registered')
       const exportFunctionPromise = registerExportFunctions(exportApiIdentifier).then(() =>
         console.log(greenHook, 'Export functions registered'),
       )
