+++ app/src/instrumentation/registerGeneralizationFunctions.ts
@@ -3,6 +3,25 @@ import { InteracitvityConfiguartion } from '@/src/app/regionen/[regionSlug]/_map
 import { TableId } from '@/src/app/regionen/[regionSlug]/_mapData/mapDataSources/tables.const'
 import { geoDataClient } from '@/src/prisma-client'
 
+async function registerCustomFunction() {
+  return geoDataClient.$executeRaw`
+    CREATE OR REPLACE FUNCTION atlas_jsonb_select(json_in JSONB, keys text[])
+    RETURNS JSONB AS $$
+    DECLARE
+      json_out JSONB := '{}';
+      key text;
+    BEGIN
+      FOREACH key IN ARRAY keys
+      loop
+        IF json_in ? key THEN
+            json_out := json_out || jsonb_build_object(key, json_in -> key);
+          END IF;
+      END LOOP;
+      RETURN json_out;
+    END;
+    $$ LANGUAGE plpgsql;`
+}
+
 async function createTileSpecification(tableName: TableId) {
   // Get column names and types
   const columnInformation = await geoDataClient.$queryRawUnsafe(`
@@ -39,6 +58,7 @@ export const SIMPLIFY_MAX_ZOOM = 14
 export async function registerGeneralizationFunctions(
   interacitvityConfiguartion: InteracitvityConfiguartion,
 ) {
+  await registerCustomFunction()
   return Promise.all(
     Object.entries(interacitvityConfiguartion).map(
       async ([tableName, { minzoom, stylingKeys }]) => {
